+++ src/main/kotlin/racinggame/validator/InputValidator.kt
@@ -1,4 +1,4 @@
-package racinggame.util
+package racinggame.validator
 
 interface InputValidator<T> {
     fun validate(input: T): T
