+++ src/main/kotlin/racinggame/validator/NumberRangeValidator.kt
@@ -1,4 +1,4 @@
-package racinggame.util
+package racinggame.validator
 
 object NumberRangeValidator : InputValidator<Int> {
     private const val INVALID_USER_INPUT_MESSAGE = "반드시 0 이상의 값이 입력되어야 합니다."
